Add definitions for symbols used in static inline functions
Several macros that used these symbols were converted to static
inline functions in a48931cc2d. However, these are not linked in
to a couple programs that include bfd, nameley readelf and elfedit.

To fix the build with compilers that always emit code for static
inline functions, add dummy definitions to those programs.